Renaming folder End to GameOver, and projectUtils.js to helpers.js
